requirements.txt
pyjwt==2.1.0
requests==2.26.0
stomp.py==7.0.0